+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version-step/utils/__tests__/remove-step.spec.ts
@@ -1,4 +1,4 @@
-import { removeStep } from 'src/modules/workflow/workflow-builder/workflow-step/utils/remove-step';
+import { removeStep } from 'src/modules/workflow/workflow-builder/workflow-version-step/utils/remove-step';
 import {
   WorkflowAction,
   WorkflowActionType,
