+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version-step/utils/remove-step.ts
